runpod_backend/requirements.txt
# pydantic==2.9.2
# pydantic-core==2.23.4
# pydantic-settings==2.5.2

# # Audio processing
# numpy==1.26.4
# soundfile==0.12.1
# librosa==0.10.1
# ffmpeg-python==0.2.0

# # LangChain
# langchain==0.3.2
# langchain-community==0.3.1
# langchain-openai==0.2.2
# langchain-teddynote==0.0.37
# langchain-huggingface==0.1.2

# # YouTube
# # pytubefix==7.1.3
# yt-dlp==2025.2.19

# # RunPod
# runpod==1.7.2

# # Utilities
# tiktoken==0.8.0
# tqdm==4.66.5

# # CUDA/GPU dependencies
# torch==2.4.1
# torchaudio==2.4.1
# faiss-gpu-cu12==1.8.0

# # HTTP client
# requests==2.31.0
# urllib3==2.2.3

# # faster-Whisper
# ctranslate2==4.4.0
# tokenizers==0.20.1
# pyannote.audio==3.3.2
# PyTorch 전용 인덱스 설정
--index-url https://download.pytorch.org/whl/cu121
torch==2.4.1
torchaudio==2.4.1

# 기본 PyPI 인덱스로 변경
--index-url https://pypi.org/simple
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.5.2

# Audio processing
numpy==1.26.4
soundfile==0.12.1
librosa==0.10.1
ffmpeg-python==0.2.0

# LangChain
langchain==0.3.18
langchain-community==0.3.18
langchain-openai==0.2.14
langchain-teddynote==0.0.37
langchain-huggingface==0.1.2

# YouTube
yt-dlp==2025.2.19

# RunPod
runpod==1.7.2

# Utilities
tiktoken==0.8.0
tqdm==4.66.5

# CUDA/GPU dependencies
faiss-gpu-cu12==1.8.0

# HTTP client
requests==2.31.0
urllib3==2.2.3

# faster-Whisper
ctranslate2==4.4.0
tokenizers==0.20.1
pyannote.audio==3.3.2
faster-whisper @ https://github.com/SYSTRAN/faster-whisper/archive/refs/heads/master.tar.gz
